chore: split by fetch call
