+++ src/lib/structures/HolodexClient.ts
@@ -20,18 +20,37 @@ export class HolodexClient {
 		return this.fetch<Holodex.Channel[]>(url, this.apiKey);
 	}
 
-	public async getLiveVideos(query: { channels?: string[]; maxUpcoming?: number; onlyUpcoming?: boolean }): Promise<Holodex.VideoWithChannel[]> {
+	public async getLiveVideos(query: { channels?: Set<string>; maxUpcoming?: number }): Promise<Holodex.VideoWithChannel[]> {
 		const { channels, maxUpcoming } = query;
 
 		const url = new URL(`${HOLODEX_BASE_URL}/users/live`);
-		if (channels) url.searchParams.append('channels', channels.toString());
+		if (channels) url.searchParams.append('channels', Array.from(channels).toString());
 
 		const result = await this.fetch<Holodex.VideoWithChannel[]>(url, this.apiKey);
 		if (maxUpcoming) {
 			return result.filter(({ channel, available_at }) => {
 				if (channels) {
-					return new Date(available_at).getTime() <= Date.now() + maxUpcoming && channels.includes(channel.id);
+					return new Date(available_at).getTime() <= Date.now() + maxUpcoming && channels.has(channel.id);
 				}
+				return new Date(available_at).getTime() <= Date.now() + maxUpcoming;
+			});
+		}
+
+		return result;
+	}
+
+	public async getVideos(query: { maxUpcoming?: number }): Promise<Holodex.VideoWithChannel[]> {
+		const { maxUpcoming } = query;
+
+		const url = new URL(`${HOLODEX_BASE_URL}/live`);
+		url.searchParams.append('status', 'upcoming,live');
+		url.searchParams.append('type', 'stream');
+		url.searchParams.append('limit', '9999');
+		url.searchParams.append('max_upcoming_hours', '1');
+
+		const result = await this.fetch<Holodex.VideoWithChannel[]>(url, this.apiKey);
+		if (maxUpcoming) {
+			return result.filter(({ available_at }) => {
 				return new Date(available_at).getTime() <= Date.now() + maxUpcoming;
 			});
 		}
